# scratch/extracted_backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: "prueba API"
  description: "Backend Spring Boot generado desde Collab Modeler (revisión 186)"
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Servidor local
security:
  - BearerAuth: []
paths:
  /api/auth/register:
    post:
      tags: [Autenticación]
      summary: Registro de usuario (primer usuario obtiene ADMIN)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '200':
          description: Usuario registrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Datos inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags: [Autenticación]
      summary: Inicio de sesión con correo y contraseña
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Credenciales válidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Credenciales inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/me:
    get:
      tags: [Autenticación]
      summary: Obtener perfil del usuario autenticado
      responses:
        '200':
          description: Perfil actual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserResponse'
        '401':
          description: No autenticado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRegisterRequest:
      type: object
      required: [email, password, fullName]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
          minLength: 6
        fullName:
          type: string
    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
...
    AuthUserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        fullName:
          type: string
        role:
          type: string
          enum: [ROLE_ADMIN, ROLE_USER]
    AuthTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        user:
          $ref: '#/components/schemas/AuthUserResponse'
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        path:
          type: string
        timestamp:
          type: string
          format: date-time
        validationErrors:
          type: object
          additionalProperties:
            type: string
...